requirements.txt
Django==3.1
celery==4.4.7
django-cors-headers==3.5.0
djangorestframework==3.11.1
psycopg2==2.8.5
python-decouple==3.3
